refactor(react): rename feature-switched to switch-feature
