--- packages/react/modules/components/feature-switched/index.js
@@ -1,1 +1,0 @@
-export default from './feature-switched';
