+++ packages/react/modules/components/switch-feature/index.js
@@ -1,0 +1,1 @@
+export default from './switch-feature';
